django/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary>=2.9
channels
daphne
whitenoise
websockets
pyotp
Pillow
requests
getenv
web3==6.19.0
